Changed line-editor logic for mixed input and output to a tty.
Previously the line-editor remembered the cursor location at the first call to read a character.  When there was output to the tty by another thread, the first character typed by the user would cause the cursor to jump back to the remembered location, which was confusing.  Now the location of the cursor is remembered only when the user types the first character.