Added "need" parameter to read-substring and read-subu8vector.
The "need" parameter is useful when a thread needs to process a stream of characters or bytes from a port and produce some other stream.  For performance, the thread should process chunks of data at a time instead of one character or byte at a time.  However the thread should react as soon as one character or byte is available.  This can be achieved by specifying 1 as the "need".