Interim changes to the networking subsystem.
Several changes were made to streamline the network addressing scheme, to improve security, and prepare the runtime for the RPC subsystem.  An important backwards incompatible change is that the server-address: parameter of open-tcp-server no longer defaults to INADDR_ANY, because otherwise inadvertent use in a quick-and-dirty script may introduce security holes.  The default is now the loopback interface (i.e. 127.0.0.1).  To get the INADDR_ANY interface you must use server-address: "*".