Added and generalized some REPL comma commands.
The comma command ,(b expr) has been added to display the backtrace of a specific continuation or thread.  Variants of ,b (i.e. ,be and ,bed) have been added to display the environment in addition to the normal frame information.  The command ,(v expr) when expr is a thread will start a REPL in that thread by interrupting the thread.