Add 'dependent' variant of pointer foreign objects.
A new `##dependent-pointer` is added, which works like a regular
`pointer`, with the difference that it has a ___FOREIGN_DEP field
which is scanned by the garbage collector.  This field can be
read with (##foreign-dependencies f).  It starts out as the empty
list, and dependencies can be added to it with
(##register-foreign-dependency! f o).  It is on purpose that no
other operations are made convenient, although determined users
can always access the __FOREIGN_DEP field directly from C code.