Add various bignum benchmarks.
Add two benchmarks based on the Chudnovsky's formula for
computing digits of pi, with 1,000 digits (for "small" bignums)
and 100,000 digits (for "bigger" bignums).  These have a nice
distribution of bignum operations.

Add a benchmark based on the old pi.scm, but replace the old
"root" code with built-in integer-sqrt.  Compute 10,000 digits.

All these benchmarks take a fraction of a second on my machines.