Compatibility with OpenSSL version 1.1.
Version 1.1 of OpenSSL introduced several non-backwards compatible
changes that broke the building of Gambit with OpenSSL features.  This
change addresses those differences and allows Gambit to use the new
version.

Note that using SSL v2 as a minimum protocol version is no longer
supported in OpenSSL 1.1.  As a result using ssl-v2 as the minimum
version in a call to make-tls-context will raise a "Wrong TLS version"
exception.