Resolve <sys/sysctl.h> header deprecation notices on linux
This isn't perfect since it doesn't take into account _when_ this
deprecation should be handled, though narrowing down this down to a
select platform seems to be an excellent first step.

As an aside, this change probably circumvents the intention of the
deprecation notice.  Maybe this wasn't a good idea to handle just
yet.