Add vector-fold{-right}, remove string-{any|every}
The definitions of string-any and string-every in SRFI 13
are not parallel to the definitions of vector-any and
vector-every in SRFI 133.  So move the definitions of vect-any
and vect-every in _std#.scm from a block executed both for
vectors and strings to a block executed just for vectors.
Put definitions of vect-fold and vect-fold-right in a block
that is executed only for vectors, not strings.

string-any and string-every are still available in SRFI 13.