SRFI 231: Safer conversions from (nested) lists and vectors to arrays.
Storage classes can be user defined, and their setters and checkers
can be general procedures.  In particular, it is (barely) conceivable
that they could modify the list or vector arguments passed to
list->array, list*->array, vector->array, and vector*-array.

So we carefully copy these arguments.  We know that the checkers and
setters of the predefined storage classes don't do such things, so
we avoid the copying if a predefined storage class is used.

So update generalized-arrays.scm, test-arrays.scm, and test/test.scm
appropriately.